Move `new_test_connection` to parent module
Make it more generic so that it can be used later in property tests as
well.
